Rename package taskstate to taskstatestorage
